+++ src/block-positioner.js
@@ -1,12 +1,5 @@
 "use strict";
 
-var template = [
-  "<div class='st-block-positioner__inner'>",
-  "<span class='st-block-positioner__selected-value'></span>",
-  "<select class='st-block-positioner__select'></select>",
-  "</div>"
-].join("\n");
-
 var BlockPositioner = function(block_element, mediator) {
   this.mediator = mediator;
   this.$block = block_element;
@@ -19,49 +12,16 @@ var BlockPositioner = function(block_element, mediator) {
 
 Object.assign(BlockPositioner.prototype, require('./function-bind'), require('./renderable'), {
 
-  total_blocks: 0,
-
-  bound: ['onBlockCountChange', 'onSelectChange', 'toggle', 'show', 'hide'],
+  bound: ['toggle', 'show', 'hide'],
 
   className: 'st-block-positioner',
   visibleClass: 'st-block-positioner--is-visible',
 
   initialize: function(){
-    this.$el.append(template);
-    this.$select = this.$('.st-block-positioner__select');
-
-    this.$select.on('change', this.onSelectChange);
-
-    this.mediator.on("block:countUpdate", this.onBlockCountChange);
-  },
-
-  onBlockCountChange: function(new_count) {
-    if (new_count !== this.total_blocks) {
-      this.total_blocks = new_count;
-      this.renderPositionList();
-    }
-  },
-
-  onSelectChange: function() {
-    var val = this.$select.val();
-    if (val !== 0) {
-      this.mediator.trigger(
-        "block:changePosition", this.$block, val,
-        (val === 1 ? 'before' : 'after'));
-      this.toggle();
-    }
-  },
-
-  renderPositionList: function() {
-    var inner = "<option value='0'>" + i18n.t("general:position") + "</option>";
-    for(var i = 1; i <= this.total_blocks; i++) {
-      inner += "<option value="+i+">"+i+"</option>";
-    }
-    this.$select.html(inner);
   },
 
   toggle: function() {
-    this.$select.val(0);
+    this.mediator.trigger('block-positioner-select:render', this);
     this.$el.toggleClass(this.visibleClass);
   },